CHG Pages Refactoring: Repos & Submission
