+++ web/src/navconfig.jsx
@@ -23,7 +23,7 @@ import Assignments from './Pages/Public/Assignments';
 import Profile from './Pages/Public/Profile';
 import Repos from './Pages/Public/Repos';
 import Submissions from './Pages/Public/Submissions';
-import Submission from './Pages/Public/Submission';
+import Submission from './Pages/Public/Submission/Submission';
 import Blog from './Pages/Public/Blog';
 import Visuals from './Pages/Public/Visuals/Visuals';
 import Lectures from './Pages/Public/Lectures/Lectures';
